01-graphrag-demo/requirements.txt
strands-agents>=1.27.0
boto3>=1.35.0
numpy>=1.24.0
neo4j>=5.28.0
neo4j-graphrag>=1.13.0
faiss-cpu>=1.9.0
python-dotenv>=1.0.1

# --- Local alternative (no AWS needed) ---
# sentence-transformers>=3.4.0  # Uncomment to use all-MiniLM-L6-v2 instead of Bedrock
